+++ sciety_labs/providers/opensearch/sync_providers.py
@@ -21,7 +21,7 @@ from sciety_labs.providers.opensearch.utils import (
     get_source_includes,
     get_vector_search_query
 )
-from sciety_labs.providers.semantic_scholar.semantic_scholar import (
+from sciety_labs.providers.semantic_scholar.providers import (
     SemanticScholarTitleAbstractEmbeddingVectorProvider
 )
 from sciety_labs.utils.datetime import get_utcnow
